+++ proofs/adapter/inMemory.php
@@ -1,30 +1,46 @@
 <?php
 declare(strict_types = 1);
 
-use Innmind\Filesystem\Adapter\InMemory;
-use Properties\Innmind\Filesystem\Adapter;
+use Innmind\Filesystem\{
+    Adapter,
+    Name,
+    File,
+    File\Content,
+};
+use Properties\Innmind\Filesystem\Adapter as PAdapter;
 use Innmind\BlackBox\Set;
 
 return static function() {
     yield properties(
-        'InMemory properties',
-        Adapter::properties(),
-        Set::call(InMemory::new(...)),
-    );
-    yield properties(
         'InMemory properties emulating filesystem',
-        Adapter::properties(),
-        Set::call(InMemory::emulateFilesystem(...)),
+        PAdapter::properties(),
+        Set::call(Adapter::inMemory(...)),
     );
 
-    foreach (Adapter::alwaysApplicable() as $property) {
+    foreach (PAdapter::alwaysApplicable() as $property) {
         yield property(
             $property,
-            Set::call(InMemory::new(...)),
-        )->named('InMemory');
-        yield property(
-            $property,
-            Set::call(InMemory::emulateFilesystem(...)),
+            Set::call(Adapter::inMemory(...)),
         )->named('InMemory emulating filesystem');
     }
+
+    yield test(
+        'Adding a file in a directory should not remove other files starting with the same name',
+        static function($assert) {
+            $adapter = Adapter::inMemory();
+            $property = new PAdapter\AddDirectoryFromAnotherAdapterWithFileAdded(
+                Name::of('0'),
+                File::named(
+                    '+1',
+                    Content::none(),
+                ),
+                File::named(
+                    '+',
+                    Content::none(),
+                ),
+            );
+
+            $property->ensureHeldBy($assert, $adapter);
+        },
+    );
 };