+++ src/File/Content/Line.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Filesystem\File\Content;
 
-use Innmind\Filesystem\Exception\DomainException;
 use Innmind\Immutable\Str;
 
 /**
@@ -11,11 +10,8 @@ use Innmind\Immutable\Str;
  */
 final class Line
 {
-    private Str $content;
-
-    private function __construct(Str $content)
+    private function __construct(private Str $content)
     {
-        $this->content = $content;
     }
 
     /**
@@ -24,7 +20,7 @@ final class Line
     public static function of(Str $content): self
     {
         if ($content->contains("\n")) {
-            throw new DomainException('New line delimiter should not appear in the line content');
+            throw new \DomainException('New line delimiter should not appear in the line content');
         }
 
         return new self($content);