+++ tests/DirectoryTest.php
@@ -8,7 +8,6 @@ use Innmind\Filesystem\{
     File,
     Name,
     File\Content,
-    Exception\DuplicatedFile,
 };
 use Innmind\Immutable\{
     Set,
@@ -51,8 +50,8 @@ class DirectoryTest extends TestCase
         $this->assertInstanceOf(Directory::class, $d2);
         $this->assertNotSame($d, $d2);
         $this->assertSame($d->name(), $d2->name());
-        $this->assertSame(0, $d->removed()->count());
-        $this->assertSame(0, $d2->removed()->count());
+        $this->assertSame(0, $d->removed()->size());
+        $this->assertSame(0, $d2->removed()->size());
         $this->assertFalse($d->contains($file->name()));
         $this->assertTrue($d2->contains($file->name()));
         $this->assertSame($file, $d2->get($file->name())->match(
@@ -106,8 +105,8 @@ class DirectoryTest extends TestCase
         $this->assertInstanceOf(Directory::class, $d2);
         $this->assertNotSame($d, $d2);
         $this->assertSame($d->name(), $d2->name());
-        $this->assertSame(0, $d->removed()->count());
-        $this->assertSame(1, $d2->removed()->count());
+        $this->assertSame(0, $d->removed()->size());
+        $this->assertSame(1, $d2->removed()->size());
         $this->assertSame(
             'bar',
             $d2
@@ -195,7 +194,7 @@ class DirectoryTest extends TestCase
                 FName::any(),
             )
             ->then(function($directory, $file) {
-                $this->expectException(DuplicatedFile::class);
+                $this->expectException(\LogicException::class);
                 $this->expectExceptionMessage("Same file '{$file->toString()}' found multiple times");
 
                 Directory::of(
@@ -216,7 +215,7 @@ class DirectoryTest extends TestCase
                 FName::any(),
             )
             ->then(function($directory, $file) {
-                $this->expectException(DuplicatedFile::class);
+                $this->expectException(\LogicException::class);
                 $this->expectExceptionMessage("Same file '{$file->toString()}' found multiple times");
 
                 Directory::named(